new PP build and make sure it is not ignored on Git
