+++ README.md
@@ -1,6 +1,8 @@
-# _PhotoPay_ In-BrOwSeR SDK
+# _PhotoPay_ In-browser SDK
 
-_PhotoPay_ In-BrOwSeR SDK enables you to perform scans of various payment barcodes in your web app, directly within the web browser, without the need for sending the image to servers for processing. You can integrate the SDK into your web app simply by following the instructions below and your web app will be able to scan and process data from the payment barcodes of various national standards. For list of all supported standards, check [this paragraph](#photopay_recognizers).
+[![Build Status](https://travis-ci.org/PhotoPay/photopay-web.svg?branch=master)](https://travis-ci.org/PhotoPay/photopay-web) [![npm version](https://badge.fury.io/js/%40microblink%2Fphotopay-web.svg)](https://badge.fury.io/js/%40microblink%2Fphotopay-web)
+
+_PhotoPay_ In-browser SDK enables you to perform scans of various payment barcodes in your web app, directly within the web browser, without the need for sending the image to servers for processing. You can integrate the SDK into your web app simply by following the instructions below and your web app will be able to scan and process data from the payment barcodes of various national standards. For list of all supported standards, check [this paragraph](#photopay_recognizers).
 
 Using _PhotoPay_ in your web app requires a valid license key. You can obtain a trial license key by registering to [Microblink dashboard](https://microblink.com/login). After registering, you will be able to generate a license key for your web app. The license key is bound to [fully qualified domain name](https://en.wikipedia.org/wiki/Fully_qualified_domain_name) of your web app, so please make sure you enter the correct name when asked. Also, keep in mind that if you plan to serve your web app from different domains, you will need different license keys.
