+++ src/Recognizers/BlinkBarcode/BarcodeRecognizer.ts
@@ -1,36 +1,5 @@
 import { Recognizer, RecognizerResult, RecognizerSettings, WasmSDK } from '../../MicroblinkSDK/DataStructures'
-
-/**
- * Possible formats of barcodes that can be detected. This enum will be returned
- * as part of BarcodeRecognizerResult interface.
- */
-export enum BarcodeFormat
-{
-    /** Indicates that no barcode has been detected. */
-    NONE = 0,
-    /** Indicates that QR code has been detected. */
-    QR_CODE,
-    /** Indicates that Data Matrix 2D barcode has been detected. */
-    DATA_MATRIX,
-    /** Indicates that UPC E 1D barcode has been detected. */
-    UPC_E,
-    /** Indicates that UPC A 1D barcode has been detected. */
-    UPC_A,
-    /** Indicates that EAN 8 1D barcode has been detected. */
-    EAN_8,
-    /** Indicates that EAN 13 1D barcode has been detected. */
-    EAN_13,
-    /** Indicates that Code 128 1D barcode has been detected. */
-    CODE_128,
-    /** Indicates that Code 39 1D barcode has been detected. */
-    CODE_39,
-    /** Indicates that ITF 1D barcode has been detected. */
-    ITF,
-    /** Indicates that Aztec 2D barcode has been detected. */
-    AZTEC_BARCODE,
-    /** Indicates that PDF417 2D barcode has been detected. */
-    PDF417_BARCODE
-}
+import { BarcodeData } from './BarcodeData'
 
 /**
  * A settings object that is used for configuring the BarcodeRecognizer.
@@ -121,20 +90,8 @@ export class BarcodeRecognizerSettings implements RecognizerSettings
  */
 export interface BarcodeRecognizerResult extends RecognizerResult
 {
-    /** Format of recognized barcode. */
-    readonly barcodeFormat: BarcodeFormat;
-
-    /**
-     * True if returned result is uncertain, i.e. if scanned barcode was incomplete (i.e.
-     * (has parts of it missing).
-     */
-    readonly uncertain: boolean;
-
-    /** String representation of data inside barcode. */
-    readonly stringData: string;
-
-    /** The raw bytes contained inside barcode. */
-    readonly rawBytes: Uint8Array;
+    /** Result of barcode recognition */
+    readonly barcodeData: BarcodeData;
 }
 
 /**
